refactor(slice-guard): decompose blueprint page into zone components
The 677-line monolith becomes a 66-line orchestrator plus eight focused
subcomponents and a utils module:

  slice-guard-utils.ts — statusGlyph + statusLabel helpers
  SgHeader.vue         — top rule + plate mark
  SgTitleBlock.vue     — blueprint corner title block
  SgHero.vue           — massive headline + lede (takes bed-count prop)
  SgBed.vue            — one bed cell w/ rest + hover-popup states
  SgPlanView.vue       — bed grid + SVG dimension line + status counts
  SgCallouts.vue       — 3 annotated callouts
  SgSpecs.vue          — implementation spec table
  SgRevs.vue           — revision history + source link

The page-slice-guard design thesis is preserved as-is: blueprint-blue
paper with cyan cross-hatch background, wireframe bed outlines with
hover-reveal data cards, SVG dimension line with tick arrows, absolute-
positioned corner title block. blink keyframe on the printing glyph
moves into SgBed with the glyph it animates.

No visual change: 'bun run test:visual -- --grep slice-guard' green on
all three viewports.
